Renamed extension methods to match standard, adjusted some extension method locations.
